Initial flake schemas
Flake schemas allow us to get rid of output-specific code in commands
like `nix flake [show|check|search]`. This is done by allowing flakes
to have a `schemas` attribute` that defines how to enumerate the
contents of the output (including documentation), and how to check
it.

In the future, this can be extended to support configurability of
flakes in a discoverable way.